fix trd tracklet plots, textual output of tracklet parameters when choosing to display

+++ Detectors/TRD/base/src/Tracklet.cxx
@@ -81,7 +81,7 @@ int Tracklet::getYbin() const
   return mY;
 }
 
-int Tracklet::getdY() const
+float Tracklet::getdY() const
 {
   return mdY;
 }
